backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
geoalchemy2==0.15.2
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-settings==2.5.2
alembic==1.13.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
httpx==0.27.2
numpy==1.26.4
opencv-python-headless==4.10.0.84
rasterio==1.3.11
gdal==3.9.0
geopandas==1.0.1
shapely==2.0.6
tensorflow==2.17.0
pillow==10.4.0
python-dotenv==1.0.1
